So, the sqlite3 module is turrible.
It silently ignores exceptions in adapters, and when they happen,
makes it look like there was never an adapter registered.

Anyways, in Py2, there is no timestamp method, so do this manually.

This probably can be done nicer if my brain worked, but I spent
enough brainpower figuring out the above.